changes in field names

+++ src/functions/graphql/queries/ads.mutation.js
@@ -1,7 +1,7 @@
 import { gql } from "@apollo/client"
 
 export const ADD_AD = gql`
-    mutation Add_AD($customer: String, $description: String, $expected_clicks: Int, $expected_reach: Int, $expiry: timestamptz, $image: String, $name: String, $url: String, $price: Int) {
+    mutation Add_AD($customer: String, $description: String, $expected_clicks: Int, $expected_reach: Int, $expiry: timestamptz, $image: String, $name: String, $url: String, $price: float8) {
   insert_ads_one(object: {customer: $customer, description: $description, expected_clicks: $expected_clicks, expected_reach: $expected_reach, expiry: $expiry, image: $image, name: $name, url: $url, price: $price}) {
     id
   }
